requirements.txt
crewai
crewai[tools]
openai
python-dotenv